Move chunk-like traits to a dedicated file (#369)
* Move chunk-like traits to a dedicated file

* Update chunk_traits.hpp

* Update chunk_traits.hpp
