+++ include/ddc/pdi.hpp
@@ -12,7 +12,7 @@
 
 #include <pdi.h>
 
-#include "ddc/chunk_common.hpp"
+#include "ddc/chunk_traits.hpp"
 
 namespace ddc {
